#138 Rename makeActionCreator to configureActionCreator
