+++ packages/actions/src/index.js
@@ -1,6 +1,6 @@
 export { prefixedValueMirror as makeActionTypes } from '@redux-tools/utils';
 
-export { default as makeActionCreator } from './makeActionCreator';
+export { default as configureActionCreator } from './configureActionCreator';
 export { default as makePayloadMetaActionCreator } from './makePayloadMetaActionCreator';
 export { default as makeConstantActionCreator } from './makeConstantActionCreator';
 export { default as makeSimpleActionCreator } from './makeSimpleActionCreator';
