+++ packages/actions/src/makePayloadMetaActionCreator.js
@@ -1,12 +1,12 @@
 import { __, nthArg } from 'ramda';
 
-import makeActionCreator from './makeActionCreator';
+import configureActionCreator from './configureActionCreator';
 
 /**
  * Creates an action creator with supplied type. The resulting action payload is the first arg.
  *
  * @sig String -> (a, {}) -> {type: String, payload: a, meta: {}}
  */
-const makePayloadMetaActionCreator = makeActionCreator(__, nthArg(0), nthArg(1));
+const makePayloadMetaActionCreator = configureActionCreator(__, nthArg(0), nthArg(1));
 
 export default makePayloadMetaActionCreator;
